Move S3Select classes to separate package
