+++ presto-hive-hadoop2/src/test/java/io/prestosql/plugin/hive/s3select/TestHiveFileSystemS3SelectPushdown.java
@@ -11,8 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.plugin.hive;
+package io.prestosql.plugin.hive.s3select;
 
+import io.prestosql.plugin.hive.AbstractTestHiveFileSystemS3;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Parameters;
